RavenDB-19071 more abstraction
